Updated CustomSettingsEditor to improve UI layout for modal dialog

+++ rflib/main/default/lwc/rflibCustomSettingsEditor/rflibCustomSettingsEditor.js
@@ -1,11 +1,13 @@
-import { LightningElement, api } from 'lwc';
+import { LightningElement, api, track } from 'lwc';
 import { ShowToastEvent } from 'lightning/platformShowToastEvent';
 import { createLogger } from 'c/rflibLogger';
 
 import getCustomSettingLabel from '@salesforce/apex/rflib_CustomSettingsEditorController.getCustomSettingLabel';
 import getCustomSettings from '@salesforce/apex/rflib_CustomSettingsEditorController.getCustomSettings';
+import getCustomSettingFields from '@salesforce/apex/rflib_CustomSettingsEditorController.getCustomSettingFields';
 import canUserModifyLoggerSettings from '@salesforce/apex/rflib_CustomSettingsEditorController.canUserModifyLoggerSettings';
 import deleteCustomSettingRecord from '@salesforce/apex/rflib_CustomSettingsEditorController.deleteCustomSettingRecord';
+import saveCustomSetting from '@salesforce/apex/rflib_CustomSettingsEditorController.saveCustomSetting';
 
 const logger = createLogger('Rflib_CustomSettingsEditor');
 
@@ -20,9 +22,19 @@ export default class Rflib_CustomSettingsEditor extends LightningElement {
     modalHeader = '';
     recordId;
     title;
-    fields = [];
     canModifySettings = false;
     isLoading = false;
+
+    // Confirmation Dialog Properties
+    showDeleteConfirmation = false;
+    deleteDialogTitle = 'Confirm Deletion';
+    deleteDialogMessage = 'Are you sure you want to delete this record?';
+    deleteDialogConfirmLabel = 'Delete';
+    deleteDialogCancelLabel = 'Cancel';
+
+    // Field Metadata and Record Values
+    fieldInfos = [];
+    @track recordValues = {};
 
     connectedCallback() {
         logger.info('Connected callback invoked. Checking user permissions and loading custom settings.');
@@ -62,20 +74,21 @@ export default class Rflib_CustomSettingsEditor extends LightningElement {
         logger.info('Loading custom settings for API name: {0}', this.customSettingsApiName);
         getCustomSettings({ customSettingsApiName: this.customSettingsApiName })
             .then((result) => {
-                // Ensure data is formatted correctly for LWC datatable
+                // Map the data for the datatable
                 this.customSettingsData = result.map((setting) => {
                     let row = {
                         id: setting.id,
-                        setupOwnerType:
-                            setting.setupOwnerType === 'Name' ? 'Default Organization Values' : setting.setupOwnerType,
-                        setupOwnerName:
-                            setting.setupOwnerType === 'Organization' ? 'Organization' : setting.setupOwnerName
+                        setupOwnerType: setting.setupOwnerType,
+                        setupOwnerName: setting.setupOwnerName
                     };
                     // Flatten the fields map to be direct properties of the row object
                     const fieldKeys = Object.keys(setting.fields);
                     fieldKeys.forEach((key) => {
                         row[key] = setting.fields[key];
                     });
+
+                    // Log the setupOwnerType for each row
+                    logger.info('Row setupOwnerType: {0}', row.setupOwnerType);
 
                     return row;
                 });
@@ -84,7 +97,7 @@ export default class Rflib_CustomSettingsEditor extends LightningElement {
                 this.columns = this.createColumns(result[0]?.fields || [], result[0]?.fieldLabels || {});
 
                 this.isLoading = false;
-                logger.info('Custom settings loaded successfully: {0}', JSON.stringify(result));
+                logger.info('Custom settings loaded successfully.');
             })
             .catch((error) => {
                 logger.error('Failed to load custom settings', error);
@@ -126,9 +139,7 @@ export default class Rflib_CustomSettingsEditor extends LightningElement {
         logger.info('Getting row actions for row: {0}', JSON.stringify(row));
         const actions = [];
         actions.push({ label: 'Edit', name: 'edit' });
-        if (row.setupOwnerType !== 'Organization') {
-            actions.push({ label: 'Delete', name: 'delete' });
-        }
+        actions.push({ label: 'Delete', name: 'delete' });
         doneCallback(actions);
     }
 
@@ -139,7 +150,7 @@ export default class Rflib_CustomSettingsEditor extends LightningElement {
         logger.info('Handling row action. Action: {0}, Record ID: {1}', actionName, this.recordId);
         switch (actionName) {
             case 'edit':
-                this.handleEditRecord();
+                this.handleEditRecord(row);
                 break;
             case 'delete':
                 this.handleDeleteRecord();
@@ -152,14 +163,71 @@ export default class Rflib_CustomSettingsEditor extends LightningElement {
     handleNewRecord() {
         logger.info('Handling new record creation.');
         this.modalHeader = 'New Custom Setting';
-        this.showModal = true;
         this.recordId = null;
+        this.recordValues = {}; // Initialize recordValues
+        this.loadFieldInfos()
+            .then(() => {
+                this.showModal = true;
+            })
+            .catch((error) => {
+                logger.error('Failed to load field infos', error);
+                this.showToast('Error', 'Failed to load field information', 'error');
+            });
     }
 
-    handleEditRecord() {
+    handleEditRecord(row) {
         logger.info('Handling edit record. Record ID: {0}', this.recordId);
         this.modalHeader = 'Edit Custom Setting';
-        this.showModal = true;
+        // Set recordValues based on the row data
+        this.recordValues = { ...row };
+        this.loadFieldInfos()
+            .then(() => {
+                this.showModal = true;
+            })
+            .catch((error) => {
+                logger.error('Failed to load field infos', error);
+                this.showToast('Error', 'Failed to load field information', 'error');
+            });
+    }
+
+    loadFieldInfos() {
+        return getCustomSettingFields({ customSettingsApiName: this.customSettingsApiName })
+            .then((fields) => {
+                // Enhance each fieldInfo object with field type flags and value
+                this.fieldInfos = fields.map((fieldInfo) => {
+                    const dataType = fieldInfo.dataType;
+                    return {
+                        ...fieldInfo,
+                        isTextField:
+                            dataType === 'STRING' || dataType === 'EMAIL' || dataType === 'PHONE' || dataType === 'URL',
+                        isNumberField: dataType === 'DOUBLE' || dataType === 'INTEGER' || dataType === 'CURRENCY',
+                        isBooleanField: dataType === 'BOOLEAN',
+                        value: this.recordValues[fieldInfo.apiName] || '' // Set the value
+                    };
+                });
+                logger.info('Field infos loaded: ' + JSON.stringify(this.fieldInfos));
+            })
+            .catch((error) => {
+                logger.error('Failed to get custom setting fields', error);
+                throw error;
+            });
+    }
+
+    handleFieldChange(event) {
+        const fieldName = event.target.dataset.fieldName;
+        let value;
+        if (event.target.type === 'checkbox') {
+            value = event.target.checked;
+        } else {
+            value = event.target.value;
+        }
+        logger.info('Field changed: {0} = {1}', fieldName, value);
+        this.recordValues[fieldName] = value;
+        // Update the value in fieldInfos
+        const fieldInfo = this.fieldInfos.find((field) => field.apiName === fieldName);
+        if (fieldInfo) {
+            fieldInfo.value = value;
+        }
     }
 
     closeModal() {
@@ -169,40 +237,58 @@ export default class Rflib_CustomSettingsEditor extends LightningElement {
 
     handleModalSave() {
         logger.info('Saving custom setting.');
-        const form = this.template.querySelector('lightning-record-edit-form');
-        if (form) {
-            form.submit();
-        }
-    }
+        // Create an object with the custom setting fields
+        const customSettingRecord = {
+            sobjectType: this.customSettingsApiName,
+            Id: this.recordId
+        };
+        this.fieldInfos.forEach((fieldInfo) => {
+            if (fieldInfo.isCreateable || fieldInfo.isUpdateable) {
+                customSettingRecord[fieldInfo.apiName] = fieldInfo.value;
+            }
+        });
+        logger.info('Custom setting record to save: {0}', JSON.stringify(customSettingRecord));
 
-    handleModalSubmit(event) {
-        event.preventDefault(); // Prevent default submission
-        const fields = event.detail.fields;
-        this.template.querySelector('lightning-record-edit-form').submit(fields);
-    }
-
-    handleModalSaveSuccess() {
-        logger.info('Custom setting saved successfully.');
-        this.showToast('Success', 'Record saved successfully.', 'success');
-        this.loadCustomSettings();
-        this.showModal = false;
+        saveCustomSetting({ customSettingRecord })
+            .then(() => {
+                logger.info('Custom setting saved successfully.');
+                this.showToast('Success', 'Record saved successfully.', 'success');
+                this.loadCustomSettings();
+                this.showModal = false;
+            })
+            .catch((error) => {
+                logger.error('Failed to save custom setting', error);
+                this.showToast('Error', 'Failed to save record', 'error');
+            });
     }
 
     handleDeleteRecord() {
-        logger.info('Handling delete record. Record ID: {0}', this.recordId);
-        deleteCustomSettingRecord({
-            customSettingsApiName: this.customSettingsApiName,
-            recordId: this.recordId
-        })
-            .then(() => {
-                logger.info('Record deleted successfully. Record ID: {0}', this.recordId);
-                this.showToast('Success', 'Record deleted successfully.', 'success');
-                this.loadCustomSettings();
+        logger.info('Preparing to delete record. Record ID: {0}', this.recordId);
+        // Set up the confirmation dialog properties
+        this.showDeleteConfirmation = true;
+    }
+
+    handleModalAction(event) {
+        const status = event.detail.status;
+        this.showDeleteConfirmation = false;
+        if (status === 'confirm') {
+            logger.info('User confirmed deletion.');
+            deleteCustomSettingRecord({
+                customSettingsApiName: this.customSettingsApiName,
+                recordId: this.recordId
             })
-            .catch((error) => {
-                logger.error('Failed to delete record: ' + JSON.stringify(error));
-                this.showToast('Error', 'Failed to delete record', 'error');
-            });
+                .then(() => {
+                    logger.info('Record deleted successfully. Record ID: {0}', this.recordId);
+                    this.showToast('Success', 'Record deleted successfully.', 'success');
+                    this.loadCustomSettings();
+                })
+                .catch((error) => {
+                    logger.error('Failed to delete record: ' + JSON.stringify(error));
+                    this.showToast('Error', 'Failed to delete record', 'error');
+                });
+        } else {
+            logger.info('User cancelled deletion.');
+        }
     }
 
     handleRefresh() {
@@ -218,5 +304,12 @@ export default class Rflib_CustomSettingsEditor extends LightningElement {
             variant: variant
         });
         this.dispatchEvent(evt);
+    }
+
+    getPicklistOptions(fieldInfo) {
+        return fieldInfo.picklistValues.map((entry) => ({
+            label: entry.value,
+            value: entry.value
+        }));
     }
 }
